Add 'sample' code provided by @chrisastaneh in chrisastaneh/ReactTraining.
I merged the ESLint config with what we already had, omitting any JQuery-related
stuff, since I expect any JQuery code in the repository is near the top of the
list of code to modernize. I also changed the name and description to refer to
this project rather than the LMS.

This fixes #2.